[chore]: split a lot of the stuff into a abstract class to reduce duplicate methods
